refactor(dinero): Use hardcoded exponent instead of hardcoded factor
When we offer better support for ISO 4217, proper exponent needs to be passed instead of assuming
all currencies have the same number of sub-units in a unit (or even have sub-units at all). This
prepares this change by using an exponent to deduct the factor to multiply or divide by.