refactor(exec): resolve/reject exit code
The previous behavior was to reject an Error if the command executed
returned a non zero exit code. Though all the stderr output was already
passed through, and the rejected error wasn't handled so it caused an
unhandled promise rejection, which creates unhelpful noise in the build
output log.

Modified the JSDoc comment to follow the ESLint rule
jsdoc/require-description-complete-sentence